+++ client/web/src/enterprise/insights/pages/insights/creation/lang-stats/LangStatsInsightCreationPage.tsx
@@ -1,7 +1,6 @@
 import classnames from 'classnames'
 import React, { useCallback, useEffect } from 'react'
 
-import { SettingsCascadeProps } from '@sourcegraph/shared/src/settings/settings'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
 import { asError } from '@sourcegraph/shared/src/util/errors'
 import { useLocalStorage } from '@sourcegraph/shared/src/util/useLocalStorage'
@@ -10,7 +9,7 @@ import { Page } from '../../../../../../components/Page'
 import { PageTitle } from '../../../../../../components/PageTitle'
 import { FORM_ERROR, FormChangeEvent } from '../../../../components/form/hooks/useForm'
 import { LangStatsInsight } from '../../../../core/types'
-import { useInsightSubjects } from '../../../../hooks/use-insight-subjects/use-insight-subjects'
+import { SupportedInsightSubject } from '../../../../core/types/subjects'
 
 import {
     LangStatsInsightCreationContent,
@@ -20,18 +19,21 @@ import styles from './LangStatsInsightCreationPage.module.scss'
 import { LangStatsCreationFormFields } from './types'
 import { getSanitizedLangStatsInsight } from './utils/insight-sanitizer'
 
-const DEFAULT_FINAL_SETTINGS = {}
-
 export interface InsightCreateEvent {
-    subjectId: string
     insight: LangStatsInsight
 }
 
-export interface LangStatsInsightCreationPageProps extends SettingsCascadeProps, TelemetryProps {
+export interface LangStatsInsightCreationPageProps extends TelemetryProps {
     /**
      * Set initial value for insight visibility setting.
      */
     visibility: string
+
+    /**
+     * List of all supported by code insights subjects that can store insight entities
+     * it's used for visibility setting section.
+     */
+    subjects: SupportedInsightSubject[]
 
     /**
      * Whenever the user submit form and clicks on save/submit button
@@ -53,16 +55,8 @@ export interface LangStatsInsightCreationPageProps extends SettingsCascadeProps,
 }
 
 export const LangStatsInsightCreationPage: React.FunctionComponent<LangStatsInsightCreationPageProps> = props => {
-    const {
-        visibility,
-        settingsCascade,
-        telemetryService,
-        onInsightCreateRequest,
-        onCancel,
-        onSuccessfulCreation,
-    } = props
+    const { visibility, subjects, telemetryService, onInsightCreateRequest, onCancel, onSuccessfulCreation } = props
 
-    const insightSubjects = useInsightSubjects({ settingsCascade })
     const [initialFormValues, setInitialFormValues] = useLocalStorage<LangStatsCreationFormFields | undefined>(
         'insights.code-stats-creation-ui',
         undefined
@@ -77,15 +71,10 @@ export const LangStatsInsightCreationPage: React.FunctionComponent<LangStatsInsi
 
     const handleSubmit = useCallback<LangStatsInsightCreationContentProps['onSubmit']>(
         async values => {
-            const subjectID = values.visibility
-
             try {
                 const insight = getSanitizedLangStatsInsight(values)
 
-                await onInsightCreateRequest({
-                    subjectId: subjectID,
-                    insight,
-                })
+                await onInsightCreateRequest({ insight })
 
                 // Clear initial values if user successfully created search insight
                 setInitialFormValues(undefined)
@@ -135,9 +124,8 @@ export const LangStatsInsightCreationPage: React.FunctionComponent<LangStatsInsi
 
             <LangStatsInsightCreationContent
                 className="pb-5"
-                settings={settingsCascade.final ?? DEFAULT_FINAL_SETTINGS}
                 initialValues={mergedInitialValues}
-                subjects={insightSubjects}
+                subjects={subjects}
                 onSubmit={handleSubmit}
                 onCancel={handleCancel}
                 onChange={handleChange}