+++ client/web/src/enterprise/insights/pages/insights/creation/search-insight/components/search-insight-creation-content/hooks/use-insight-creation-form/use-insight-creation-form.ts
@@ -1,8 +1,7 @@
-import { Settings } from '../../../../../../../../../../schema/settings.schema'
+import { useAsyncInsightTitleValidator } from '../../../../../../../../components/form/hooks/use-async-insight-title-validator'
 import { useField, useFieldAPI } from '../../../../../../../../components/form/hooks/useField'
 import { Form, FormChangeEvent, SubmissionErrors, useForm } from '../../../../../../../../components/form/hooks/useForm'
-import { useInsightTitleValidator } from '../../../../../../../../components/form/hooks/useInsightTitleValidator'
-import { InsightTypePrefix } from '../../../../../../../../core/types'
+import { createRequiredValidator } from '../../../../../../../../components/form/validators'
 import { isUserSubject, SupportedInsightSubject } from '../../../../../../../../core/types/subjects'
 import { CreateInsightFormFields, EditableDataSeries, InsightStep } from '../../../../types'
 import { INITIAL_INSIGHT_VALUES } from '../../initial-insight-values'
@@ -13,41 +12,18 @@ import {
     seriesRequired,
 } from '../../validators'
 
+const titleRequiredValidator = createRequiredValidator('Title is a required field.')
+
 export interface UseInsightCreationFormProps {
-    /**
-     * Final (merged) settings cascade  objects
-     */
-    settings?: Settings | null
-
-    /**
-     * List of all supportable insight subjects
-     */
+    mode: 'creation' | 'edit'
     subjects?: SupportedInsightSubject[]
-
-    /**
-     * Initial value for all form fields
-     */
     initialValue?: Partial<CreateInsightFormFields>
-
-    /**
-     * Set initial touched state for all form fields.
-     */
-    touched?: boolean
-
-    /**
-     * Submit handler for form element.
-     */
     onSubmit: (values: CreateInsightFormFields) => SubmissionErrors | Promise<SubmissionErrors> | void
-
-    /**
-     * Change handlers is called every time when user changed any field within the form.
-     */
     onChange?: (event: FormChangeEvent<CreateInsightFormFields>) => void
 }
 
 export interface InsightCreationForm {
     form: Form<CreateInsightFormFields>
-
     title: useFieldAPI<string>
     repositories: useFieldAPI<string>
     visibility: useFieldAPI<string>
@@ -58,13 +34,15 @@ export interface InsightCreationForm {
 }
 
 /**
- * Hooks absorbs all insight creation form logic (field state managements, validations, fields dependencies)
+ * Hooks absorbs all insight creation form logic (field state managements,
+ * validations, fields dependencies)
  */
 export function useInsightCreationForm(props: UseInsightCreationFormProps): InsightCreationForm {
-    const { subjects = [], initialValue = {}, touched, settings, onSubmit, onChange } = props
+    const { mode, subjects = [], initialValue = {}, onSubmit, onChange } = props
 
     // Calculate initial value for visibility settings
     const userSubjectID = subjects.find(isUserSubject)?.id ?? ''
+    const isEdit = mode === 'edit'
 
     const form = useForm<CreateInsightFormFields>({
         initialValues: {
@@ -74,7 +52,7 @@ export function useInsightCreationForm(props: UseInsightCreationFormProps): Insi
         },
         onSubmit,
         onChange,
-        touched,
+        touched: isEdit,
     })
 
     const allReposMode = useField({
@@ -91,13 +69,15 @@ export function useInsightCreationForm(props: UseInsightCreationFormProps): Insi
     })
 
     const isAllReposMode = allReposMode.input.value
+    const asyncTitleValidator = useAsyncInsightTitleValidator({
+        mode,
+        initialTitle: form.formAPI.initialValues.title,
+    })
 
-    // We can't have two or more insights with the same name, since we rely on name as on id of insights.
-    const titleValidator = useInsightTitleValidator({ settings, insightType: InsightTypePrefix.search })
     const title = useField({
         name: 'title',
         formApi: form.formAPI,
-        validators: { sync: titleValidator },
+        validators: { sync: titleRequiredValidator, async: asyncTitleValidator },
     })
 
     const repositories = useField({