+++ client/web/src/search/panels/HomePanels.tsx
@@ -1,3 +1,4 @@
+import classNames from 'classnames'
 import * as React from 'react'
 
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
@@ -6,6 +7,7 @@ import { HomePanelsProps, PatternTypeProps } from '..'
 import { AuthenticatedUser } from '../../auth'
 
 import { CommunitySearchContextsPanel } from './CommunitySearchContextPanel'
+import styles from './HomePanels.module.scss'
 import { RecentFilesPanel } from './RecentFilesPanel'
 import { RecentSearchesPanel } from './RecentSearchesPanel'
 import { RepositoriesPanel } from './RepositoriesPanel'
@@ -17,18 +19,18 @@ interface Props extends Pick<PatternTypeProps, 'patternType'>, TelemetryProps, H
 }
 
 export const HomePanels: React.FunctionComponent<Props> = (props: Props) => (
-    <div className="home-panels container">
+    <div className={classNames('container', styles.homePanels)} data-testid="home-panels">
         <div className="row">
-            <RepositoriesPanel {...props} className="home-panels__panel col-lg-4" />
-            <RecentSearchesPanel {...props} className="home-panels__panel col-lg-8" />
+            <RepositoriesPanel {...props} className={classNames('col-lg-4', styles.panel)} />
+            <RecentSearchesPanel {...props} className={classNames('col-lg-8', styles.panel)} />
         </div>
         <div className="row">
-            <RecentFilesPanel {...props} className="home-panels__panel col-lg-7" />
+            <RecentFilesPanel {...props} className={classNames('col-lg-7', styles.panel)} />
 
             {props.isSourcegraphDotCom ? (
-                <CommunitySearchContextsPanel {...props} className="home-panels__panel col-lg-5" />
+                <CommunitySearchContextsPanel {...props} className={classNames('col-lg-5', styles.panel)} />
             ) : (
-                <SavedSearchesPanel {...props} className="home-panels__panel col-lg-5" />
+                <SavedSearchesPanel {...props} className={classNames('col-lg-5', styles.panel)} />
             )}
         </div>
     </div>