+++ enterprise/cmd/precise-code-intel-worker/internal/worker/mock_iface_test.go
@@ -2138,6 +2138,10 @@ type MockLSIFStore struct {
 	// object controlling the behavior of the method
 	// WriteDocumentationPathInfo.
 	WriteDocumentationPathInfoFunc *LSIFStoreWriteDocumentationPathInfoFunc
+	// WriteDocumentationSearchPreworkFunc is an instance of a mock function
+	// object controlling the behavior of the method
+	// WriteDocumentationSearchPrework.
+	WriteDocumentationSearchPreworkFunc *LSIFStoreWriteDocumentationSearchPreworkFunc
 	// WriteDocumentsFunc is an instance of a mock function object
 	// controlling the behavior of the method WriteDocuments.
 	WriteDocumentsFunc *LSIFStoreWriteDocumentsFunc
@@ -2177,13 +2181,18 @@ func NewMockLSIFStore() *MockLSIFStore {
 			},
 		},
 		WriteDocumentationPagesFunc: &LSIFStoreWriteDocumentationPagesFunc{
-			defaultHook: func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData) error {
+			defaultHook: func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData, int, int) error {
 				return nil
 			},
 		},
 		WriteDocumentationPathInfoFunc: &LSIFStoreWriteDocumentationPathInfoFunc{
 			defaultHook: func(context.Context, int, chan *precise.DocumentationPathInfoData) error {
 				return nil
+			},
+		},
+		WriteDocumentationSearchPreworkFunc: &LSIFStoreWriteDocumentationSearchPreworkFunc{
+			defaultHook: func(context.Context, dbstore.Upload, *types.Repo, bool) (int, int, error) {
+				return 0, 0, nil
 			},
 		},
 		WriteDocumentsFunc: &LSIFStoreWriteDocumentsFunc{
@@ -2230,6 +2239,9 @@ func NewMockLSIFStoreFrom(i LSIFStore) *MockLSIFStore {
 		},
 		WriteDocumentationPathInfoFunc: &LSIFStoreWriteDocumentationPathInfoFunc{
 			defaultHook: i.WriteDocumentationPathInfo,
+		},
+		WriteDocumentationSearchPreworkFunc: &LSIFStoreWriteDocumentationSearchPreworkFunc{
+			defaultHook: i.WriteDocumentationSearchPrework,
 		},
 		WriteDocumentsFunc: &LSIFStoreWriteDocumentsFunc{
 			defaultHook: i.WriteDocuments,
@@ -2678,24 +2690,24 @@ func (c LSIFStoreWriteDocumentationMappingsFuncCall) Results() []interface{} {
 // WriteDocumentationPages method of the parent MockLSIFStore instance is
 // invoked.
 type LSIFStoreWriteDocumentationPagesFunc struct {
-	defaultHook func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData) error
-	hooks       []func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData) error
+	defaultHook func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData, int, int) error
+	hooks       []func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData, int, int) error
 	history     []LSIFStoreWriteDocumentationPagesFuncCall
 	mutex       sync.Mutex
 }
 
 // WriteDocumentationPages delegates to the next hook function in the queue
 // and stores the parameter and result values of this invocation.
-func (m *MockLSIFStore) WriteDocumentationPages(v0 context.Context, v1 dbstore.Upload, v2 *types.Repo, v3 bool, v4 chan *precise.DocumentationPageData) error {
-	r0 := m.WriteDocumentationPagesFunc.nextHook()(v0, v1, v2, v3, v4)
-	m.WriteDocumentationPagesFunc.appendCall(LSIFStoreWriteDocumentationPagesFuncCall{v0, v1, v2, v3, v4, r0})
+func (m *MockLSIFStore) WriteDocumentationPages(v0 context.Context, v1 dbstore.Upload, v2 *types.Repo, v3 bool, v4 chan *precise.DocumentationPageData, v5 int, v6 int) error {
+	r0 := m.WriteDocumentationPagesFunc.nextHook()(v0, v1, v2, v3, v4, v5, v6)
+	m.WriteDocumentationPagesFunc.appendCall(LSIFStoreWriteDocumentationPagesFuncCall{v0, v1, v2, v3, v4, v5, v6, r0})
 	return r0
 }
 
 // SetDefaultHook sets function that is called when the
 // WriteDocumentationPages method of the parent MockLSIFStore instance is
 // invoked and the hook queue is empty.
-func (f *LSIFStoreWriteDocumentationPagesFunc) SetDefaultHook(hook func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData) error) {
+func (f *LSIFStoreWriteDocumentationPagesFunc) SetDefaultHook(hook func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData, int, int) error) {
 	f.defaultHook = hook
 }
 
@@ -2704,7 +2716,7 @@ func (f *LSIFStoreWriteDocumentationPagesFunc) SetDefaultHook(hook func(context.
 // invokes the hook at the front of the queue and discards it. After the
 // queue is empty, the default hook function is invoked for any future
 // action.
-func (f *LSIFStoreWriteDocumentationPagesFunc) PushHook(hook func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData) error) {
+func (f *LSIFStoreWriteDocumentationPagesFunc) PushHook(hook func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData, int, int) error) {
 	f.mutex.Lock()
 	f.hooks = append(f.hooks, hook)
 	f.mutex.Unlock()
@@ -2713,7 +2725,7 @@ func (f *LSIFStoreWriteDocumentationPagesFunc) PushHook(hook func(context.Contex
 // SetDefaultReturn calls SetDefaultDefaultHook with a function that returns
 // the given values.
 func (f *LSIFStoreWriteDocumentationPagesFunc) SetDefaultReturn(r0 error) {
-	f.SetDefaultHook(func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData) error {
+	f.SetDefaultHook(func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData, int, int) error {
 		return r0
 	})
 }
@@ -2721,12 +2733,12 @@ func (f *LSIFStoreWriteDocumentationPagesFunc) SetDefaultReturn(r0 error) {
 // PushReturn calls PushDefaultHook with a function that returns the given
 // values.
 func (f *LSIFStoreWriteDocumentationPagesFunc) PushReturn(r0 error) {
-	f.PushHook(func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData) error {
+	f.PushHook(func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData, int, int) error {
 		return r0
 	})
 }
 
-func (f *LSIFStoreWriteDocumentationPagesFunc) nextHook() func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData) error {
+func (f *LSIFStoreWriteDocumentationPagesFunc) nextHook() func(context.Context, dbstore.Upload, *types.Repo, bool, chan *precise.DocumentationPageData, int, int) error {
 	f.mutex.Lock()
 	defer f.mutex.Unlock()
 
@@ -2775,6 +2787,12 @@ type LSIFStoreWriteDocumentationPagesFuncCall struct {
 	// Arg4 is the value of the 5th argument passed to this method
 	// invocation.
 	Arg4 chan *precise.DocumentationPageData
+	// Arg5 is the value of the 6th argument passed to this method
+	// invocation.
+	Arg5 int
+	// Arg6 is the value of the 7th argument passed to this method
+	// invocation.
+	Arg6 int
 	// Result0 is the value of the 1st result returned from this method
 	// invocation.
 	Result0 error
@@ -2783,7 +2801,7 @@ type LSIFStoreWriteDocumentationPagesFuncCall struct {
 // Args returns an interface slice containing the arguments of this
 // invocation.
 func (c LSIFStoreWriteDocumentationPagesFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2, c.Arg3, c.Arg4}
+	return []interface{}{c.Arg0, c.Arg1, c.Arg2, c.Arg3, c.Arg4, c.Arg5, c.Arg6}
 }
 
 // Results returns an interface slice containing the results of this
@@ -2902,6 +2920,128 @@ func (c LSIFStoreWriteDocumentationPathInfoFuncCall) Args() []interface{} {
 // invocation.
 func (c LSIFStoreWriteDocumentationPathInfoFuncCall) Results() []interface{} {
 	return []interface{}{c.Result0}
+}
+
+// LSIFStoreWriteDocumentationSearchPreworkFunc describes the behavior when
+// the WriteDocumentationSearchPrework method of the parent MockLSIFStore
+// instance is invoked.
+type LSIFStoreWriteDocumentationSearchPreworkFunc struct {
+	defaultHook func(context.Context, dbstore.Upload, *types.Repo, bool) (int, int, error)
+	hooks       []func(context.Context, dbstore.Upload, *types.Repo, bool) (int, int, error)
+	history     []LSIFStoreWriteDocumentationSearchPreworkFuncCall
+	mutex       sync.Mutex
+}
+
+// WriteDocumentationSearchPrework delegates to the next hook function in
+// the queue and stores the parameter and result values of this invocation.
+func (m *MockLSIFStore) WriteDocumentationSearchPrework(v0 context.Context, v1 dbstore.Upload, v2 *types.Repo, v3 bool) (int, int, error) {
+	r0, r1, r2 := m.WriteDocumentationSearchPreworkFunc.nextHook()(v0, v1, v2, v3)
+	m.WriteDocumentationSearchPreworkFunc.appendCall(LSIFStoreWriteDocumentationSearchPreworkFuncCall{v0, v1, v2, v3, r0, r1, r2})
+	return r0, r1, r2
+}
+
+// SetDefaultHook sets function that is called when the
+// WriteDocumentationSearchPrework method of the parent MockLSIFStore
+// instance is invoked and the hook queue is empty.
+func (f *LSIFStoreWriteDocumentationSearchPreworkFunc) SetDefaultHook(hook func(context.Context, dbstore.Upload, *types.Repo, bool) (int, int, error)) {
+	f.defaultHook = hook
+}
+
+// PushHook adds a function to the end of hook queue. Each invocation of the
+// WriteDocumentationSearchPrework method of the parent MockLSIFStore
+// instance invokes the hook at the front of the queue and discards it.
+// After the queue is empty, the default hook function is invoked for any
+// future action.
+func (f *LSIFStoreWriteDocumentationSearchPreworkFunc) PushHook(hook func(context.Context, dbstore.Upload, *types.Repo, bool) (int, int, error)) {
+	f.mutex.Lock()
+	f.hooks = append(f.hooks, hook)
+	f.mutex.Unlock()
+}
+
+// SetDefaultReturn calls SetDefaultDefaultHook with a function that returns
+// the given values.
+func (f *LSIFStoreWriteDocumentationSearchPreworkFunc) SetDefaultReturn(r0 int, r1 int, r2 error) {
+	f.SetDefaultHook(func(context.Context, dbstore.Upload, *types.Repo, bool) (int, int, error) {
+		return r0, r1, r2
+	})
+}
+
+// PushReturn calls PushDefaultHook with a function that returns the given
+// values.
+func (f *LSIFStoreWriteDocumentationSearchPreworkFunc) PushReturn(r0 int, r1 int, r2 error) {
+	f.PushHook(func(context.Context, dbstore.Upload, *types.Repo, bool) (int, int, error) {
+		return r0, r1, r2
+	})
+}
+
+func (f *LSIFStoreWriteDocumentationSearchPreworkFunc) nextHook() func(context.Context, dbstore.Upload, *types.Repo, bool) (int, int, error) {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
+
+	if len(f.hooks) == 0 {
+		return f.defaultHook
+	}
+
+	hook := f.hooks[0]
+	f.hooks = f.hooks[1:]
+	return hook
+}
+
+func (f *LSIFStoreWriteDocumentationSearchPreworkFunc) appendCall(r0 LSIFStoreWriteDocumentationSearchPreworkFuncCall) {
+	f.mutex.Lock()
+	f.history = append(f.history, r0)
+	f.mutex.Unlock()
+}
+
+// History returns a sequence of
+// LSIFStoreWriteDocumentationSearchPreworkFuncCall objects describing the
+// invocations of this function.
+func (f *LSIFStoreWriteDocumentationSearchPreworkFunc) History() []LSIFStoreWriteDocumentationSearchPreworkFuncCall {
+	f.mutex.Lock()
+	history := make([]LSIFStoreWriteDocumentationSearchPreworkFuncCall, len(f.history))
+	copy(history, f.history)
+	f.mutex.Unlock()
+
+	return history
+}
+
+// LSIFStoreWriteDocumentationSearchPreworkFuncCall is an object that
+// describes an invocation of method WriteDocumentationSearchPrework on an
+// instance of MockLSIFStore.
+type LSIFStoreWriteDocumentationSearchPreworkFuncCall struct {
+	// Arg0 is the value of the 1st argument passed to this method
+	// invocation.
+	Arg0 context.Context
+	// Arg1 is the value of the 2nd argument passed to this method
+	// invocation.
+	Arg1 dbstore.Upload
+	// Arg2 is the value of the 3rd argument passed to this method
+	// invocation.
+	Arg2 *types.Repo
+	// Arg3 is the value of the 4th argument passed to this method
+	// invocation.
+	Arg3 bool
+	// Result0 is the value of the 1st result returned from this method
+	// invocation.
+	Result0 int
+	// Result1 is the value of the 2nd result returned from this method
+	// invocation.
+	Result1 int
+	// Result2 is the value of the 3rd result returned from this method
+	// invocation.
+	Result2 error
+}
+
+// Args returns an interface slice containing the arguments of this
+// invocation.
+func (c LSIFStoreWriteDocumentationSearchPreworkFuncCall) Args() []interface{} {
+	return []interface{}{c.Arg0, c.Arg1, c.Arg2, c.Arg3}
+}
+
+// Results returns an interface slice containing the results of this
+// invocation.
+func (c LSIFStoreWriteDocumentationSearchPreworkFuncCall) Results() []interface{} {
+	return []interface{}{c.Result0, c.Result1, c.Result2}
 }
 
 // LSIFStoreWriteDocumentsFunc describes the behavior when the